pyproject.toml
[tool.poetry]
name = "game bot"
version = "0.1.0"
description = ""
authors = ["Vincent Law <vincentlaw2699@bwg-io.site>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.3"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"
aiogram = "^2.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
